refactor(codex): share usageToTurnTelemetry from harness.ts
The mapping is harness-agnostic; the next harness to report turn
telemetry imports it from the module that owns the result shape.
